refactor: section component 분리
